Bugfix/feature: attr_implement respects superclasses
For example, we used `attr_implement :foo?` in a module, but if we mixed it into an Active Record model that got 'foo?' defined by AR as a column query method, it would not be respected.

Intentionally does not define `respond_to_missing?` since this feature
is really just a glorified "undefined method" anyway.