Fix `rattr_initialize` example and typos
The example for `rattr_initialize` was creating an instance with a
single argument, while calling `rattr_initialize` with two arguments.
Running the code would cause an `ArgumentError` in `validate_arity`.
This change adds an argument to the instance initialization.

Also fixes typos in example calls to `rattr_initialize` and
`pattr_initialize`.