Pass kwargs explicitly in explicit static_facade
Beginning with Ruby 2.7.1, Ruby is a bit more strict about argument
splatting and spreading. Specifically, there is even more of a
difference between keyword arguments and positional arguments than there
was before, and when Ruby 3 is released, `*args` will no longer capture
keyword arguments — you must use `**kwargs`. ([See this post.][1]) So
when we are causing `.static_facade` to delegate most of its arguments
to `.new`, we must use a double-splat, otherwise we will get a warning.

[1]: https://www.ruby-lang.org/en/news/2019/12/12/separation-of-positional-and-keyword-arguments-in-ruby-3-0/